% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendCheckDigit.R
\name{AppendCheckDigit}
\alias{AppendCheckDigit}
\title{Append a check digit to a value}
\usage{
AppendCheckDigit(x, method)
}
\arguments{
\item{x}{character vector of values requiring a check digit}

\item{method}{a character value naming the check digit algorithm to use}
}
\value{
character vector similar to \code{x} but with a check digit appended to
each value
}
\description{
Append a check digit to a value
}
\examples{
AppendCheckDigit('1234', 'Verhoeff') # 12340

}
