% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verhoeff.R
\name{verhoeff}
\alias{verhoeff}
\alias{AppendCheckDigit.Verhoeff}
\alias{VerifyCheckDigit.Verhoeff}
\title{Verhoeff algorithm implementation}
\usage{
AppendCheckDigit.Verhoeff(x)

VerifyCheckDigit.Verhoeff(x)
}
\arguments{
\item{x}{character vector of values}
}
\description{
The Verhoeff algorithm is appropriate for numerical data and detects all
single-digit substitutions (x to y) and adjacent digit transpositions (xy to
yx). In addition, the Verhoeff check digit can detect most twin errors (xx to
yy), jump twin errors (xyx to zyz), jump transpositions (xyz to zyx), and
phonetic errors ('sixty' to 'sixteen').
}
\references{
Verhoeff, J. "Error Detecting Decimal Codes", Mathematical Centre
Tract 29, The Mathematical Centre, Amsterdam, 1969.
}
