% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDigit-package.R
\docType{package}
\name{CheckDigit-package}
\alias{CheckDigit}
\alias{CheckDigit-package}
\title{CheckDigit: Calculate and Verify Check Digits}
\description{
Check digits are used like file hashes to verify that a number has been transcribed accurately. The functions provided by this package help to calculate and verify check digits according to various algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fascinatingfingers.gitlab.io/checkdigit}
  \item \url{https://gitlab.com/fascinatingfingers/checkdigit}
  \item Report bugs at \url{https://gitlab.com/fascinatingfingers/checkdigit/-/issues}
}

}
\author{
\strong{Maintainer}: Justin Brantley \email{fascinatingfingers@icloud.com}

}
\keyword{internal}
