% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDigit-package.R
\docType{package}
\name{CheckDigit-package}
\alias{CheckDigit}
\alias{CheckDigit-package}
\title{CheckDigit: Calculate and Verify Check Digits}
\description{
The CheckDigit package provides a set of functions to calculate check digits
according to various algorithms and to verify whether a value ends in a valid
check digit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/fascinatingfingers/checkdigit}
  \item Report bugs at \url{https://gitlab.com/fascinatingfingers/checkdigit/issues}
}

}
\keyword{internal}
